move Writer to its own file
